Yet MORE hacks, I created a bug where the HP column lingers after a unit dies, so I fixed it. Also did some more optimization (remove instances of type casting in tight loops)

+++ core/src/main/java/com/vdxp/demon_front/core/map/MapTile.java
@@ -36,7 +36,7 @@ public abstract class MapTile extends Drawable implements Collidable {
         }
     }
 
-    public int getTileDistFrom(float targetX, float targetY) {
+    public float getTileDistFrom(float targetX, float targetY) {
         return (int) (Math.abs(
                   Math.sqrt(
                       (double) (
